# Conflicts:
#	upsr/impl/src/main/java/cn/org/upbnc/api/TunnelApi.java
#	upsr/impl/src/main/java/cn/org/upbnc/api/impl/TunnelApiImpl.java
#	upsr/impl/src/main/java/cn/org/upbnc/odlapi/TunnelODLApi.java
#	upsr/impl/src/main/java/cn/org/upbnc/service/TunnelService.java
#	upsr/impl/src/main/java/cn/org/upbnc/service/impl/TunnelServiceImpl.java

+++ upsr/impl/src/main/java/cn/org/upbnc/api/impl/TunnelApiImpl.java
@@ -37,6 +37,14 @@ public class TunnelApiImpl implements TunnelApi {
 
     @Override
     public Map<String, Object> createTunnel(TunnelServiceEntity tunnelServiceEntity) {
+        Map<String, Object> map = new HashMap<>();
+        map.put(ResponseEnum.CODE.getName(), CodeEnum.ERROR.getName());
+        if (null == tunnelServiceEntity.getTunnelName() || "".equals(tunnelServiceEntity.getTunnelName()) ||
+                null == tunnelServiceEntity.getTunnelId() || "".equals(tunnelServiceEntity.getTunnelId()) ||
+                null == tunnelServiceEntity.getRouterId() || "".equals(tunnelServiceEntity.getRouterId())) {
+            map.put(ResponseEnum.MESSAGE.getName(), "tunnel name ,id or routerId is null");
+            return map;
+        }
         return tunnelService.createTunnel(tunnelServiceEntity);
     }
 
@@ -47,12 +55,18 @@ public class TunnelApiImpl implements TunnelApi {
 
     @Override
     public Map<String, Object> deleteTunnel(String routerId, String tunnelName) {
+        Map<String, Object> map = new HashMap<>();
+        map.put(ResponseEnum.CODE.getName(), CodeEnum.ERROR.getName());
+        if (null == routerId || "".equals(routerId) || null == tunnelName || "".equals(tunnelName)) {
+            map.put(ResponseEnum.MESSAGE.getName(), "tunnel name or routerId is null");
+            return map;
+        }
         return tunnelService.deleteTunnel(routerId, tunnelName);
     }
 
     @Override
-    public Map<String, Object> getAllTunnel() {
-        return null;
+    public Map<String, Object> getAllTunnel(String routerId, String tunnelName) {
+        return tunnelService.getAllTunnel(routerId, tunnelName);
     }
 
     @Override
